Finish redux forms logic example
- Now it is super fast by default
- Two simple rules: Do not nest state selections and… if needed, use
redux-virtualised.
